Add Driver Binder application, support bind/unbind the host drivers
When BMC boots up with the host is Off, linux kernel will load the host
drivers, but the devices are not available so it will not bind the drivers.
Moreover, linux kernel don't support auto rebind the driver when the device
is ready. So OpenBmc have to support the application to bind/unbind the
host drivers base on the host status.
This commit adds support that application.

Tested:
	1. Boot up BMC with the host is Off.
	2. The host drivers are not binded.
	3. Turn on the host.
	4. The host drivers are binded.
	5. Turn off the host.
	6. The host drivers are unbinded.

Signed-off-by: Thu Ba Nguyen <tbnguyen@amperecomputing.com>